split actions + support add new dashboard
