+++ src/reducers/selected.js
@@ -5,7 +5,7 @@ import { validateReducer } from '../util';
 
 export const actionTypes = {
     SET_SELECTED_ID: 'SET_SELECTED_ID',
-    SET_SELECTED_EDIT: '',
+    SET_SELECTED_EDIT: 'SET_SELECTED_EDIT',
     SET_SELECTED_ISLOADING: 'SET_SELECTED_ISLOADING',
     SET_SELECTED_SHOWDESCRIPTION: 'SET_SELECTED_SHOWDESCRIPTION',
 };
